refactor(desktop): unify Desktop→Rust API to OpenAPI-generated client
- Add utoipa annotations to usage, models, ratelimits handlers
- Extract usage handlers from router.rs into handler/usage.rs
- Add typed UsageBucket, UsageSnapshot, ModelsResponse schemas
- Regenerate openapi.json (8 endpoints, 19 schemas)
- Regenerate Swift client via swift-openapi-generator
- Replace hand-written APIClient.swift with generated Client
- Update DataService to use generated client for all data fetching
- Add TypeAliases.swift bridging Components.Schemas to short names
- Adapt Views for additionalProperties and Int64 type changes
- Add 'make openapi' target to Makefile for automated regeneration
- Delete APIClient.swift; CLIRunner retained for interactive-only ops

+++ desktop/Byokey/Views/Dashboard/DashboardHistoryChart.swift
@@ -9,7 +9,7 @@ struct DashboardHistoryChart: View {
         return Dictionary(grouping: history.buckets, by: \.period_start)
             .map { ts, buckets in
                 (date: Date(timeIntervalSince1970: TimeInterval(ts)),
-                 requests: buckets.reduce(0) { $0 + $1.request_count })
+                 requests: UInt64(buckets.reduce(0) { $0 + $1.request_count }))
             }
             .sorted { $0.date < $1.date }
     }
